reorganize and increase see threshold
